refactor protected route
