+++ src/components/menu/index.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import classNames from 'classnames';
 import { Link, NavLink } from 'react-router-dom';
 
 import { signOut } from 'storage';
@@ -26,7 +27,7 @@ const Menu = ({ isAuth }: MenuProps) => {
   const itemsAuth = (
     <div className="justify-center flex flex-1 flex-col md:flex-row items-center md:items-start">
       {items.map(({ name, path }) => (
-        <div className="mx-5 my-4 md:my-0 ">
+        <div className="mx-5 my-4 md:my-0" key={name}>
           <NavLink
             key={name}
             exact
@@ -48,9 +49,10 @@ const Menu = ({ isAuth }: MenuProps) => {
             <img className="w-12" src={User} alt="" />
           </div>
           <div
-            className={`cursor-pointer p-5 transform duration-200 ${
-              isSettingOpen && 'rotate-180'
-            }`}
+            className={classNames({
+              'cursor-pointer p-5 transform duration-200': true,
+              'rotate-180': isSettingOpen,
+            })}
             onClick={() => setSetting((prev) => !prev)}
           >
             <img src={Arrow} alt="" />
@@ -58,9 +60,11 @@ const Menu = ({ isAuth }: MenuProps) => {
         </div>
 
         <div
-          className={`flex text-center flex-col md:bg-blue-500 pt-4 pb-3 w-full rounded mt-3 text-2xl md:text-lg  ${
-            !isSettingOpen && 'hidden'
-          }`}
+          className={classNames({
+            'flex text-center flex-col md:bg-blue-500 pt-4 pb-3 w-full rounded mt-3 text-2xl md:text-lg':
+              true,
+            hidden: !isSettingOpen,
+          })}
         >
           <div
             className="hover:font-bold cursor-pointer"
