+++ src/views/movies/index.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import classNames from 'classnames';
 import { Link } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 
@@ -25,7 +26,10 @@ const Movies = () => {
       <div className="container mx-auto mt-40 mb-20">
         <div className="flex w-full flex-col md:flex-row text-white flex-wrap">
           {movies?.map((movie: ItemInterface) => (
-            <div className="flex flex-col px-3 w-full md:w-1/4 mb-10 cursor-pointer">
+            <div
+              className="flex flex-col px-3 w-full md:w-1/4 mb-10 cursor-pointer"
+              key={movie.id}
+            >
               <Link to={`/trailer/${movie.id}`}>
                 <img className="h-50 w-full" src={movie.card_url} alt="" />
                 <div className="flex my-4 items-center">
@@ -40,18 +44,18 @@ const Movies = () => {
           ))}
         </div>
         <div className="flex justify-center text-center text-white">
-          {Array(pages)
-            .fill('')
-            .map((_, index) => (
-              <div
-                className={`m-1 px-2 py-1 rounded cursor-pointer ${
-                  page === index + 1 && 'bg-blue-500'
-                }`}
-                onClick={() => setPage(index + 1)}
-              >
-                {index + 1}
-              </div>
-            ))}
+          {Array.from(Array(pages).keys()).map((_, index) => (
+            <div
+              className={classNames({
+                'm-1 px-2 py-1 rounded cursor-pointer': true,
+                'bg-blue-500': page === index + 1,
+              })}
+              onClick={() => setPage(index + 1)}
+              key={_}
+            >
+              {index + 1}
+            </div>
+          ))}
         </div>
       </div>
     </div>
