+++ src/views/trailer/index.tsx
@@ -43,11 +43,9 @@ const Trailer = () => {
             <div>Quentien cuarentino</div>
           </div>
           <div className="flex items-center">
-            {Array(movie?.stars)
-              .fill('')
-              .map(() => (
-                <img className="w-4 mr-1" src={Star} alt="" />
-              ))}
+            {Array.from(Array(movie.stars).keys()).map((_) => (
+              <img className="w-4 mr-1" src={Star} alt="" key={_} />
+            ))}
           </div>
         </div>
         <div
